+++ service-abstract/src/main/java/jm/ChannelService.java
@@ -14,7 +14,7 @@ public interface ChannelService {
 
     void updateChannel(Channel channel);
 
-    Channel getChannelById(int id);
+    Channel getChannelById(Long id);
 
     Channel getChannelByName(String name);
